+++ superset-frontend/src/dashboard/components/URLShortLinkButton/URLShortLinkButton.test.tsx
@@ -19,7 +19,7 @@
 import { render, screen, userEvent } from 'spec/helpers/testing-library';
 import fetchMock from 'fetch-mock';
 import URLShortLinkButton from 'src/dashboard/components/URLShortLinkButton';
-import AntdToastContainer from 'src/components/MessageToasts/AntdToastContainer';
+import ToastContainer from 'src/components/MessageToasts/ToastContainer';
 
 const DASHBOARD_ID = 10;
 const PERMALINK_PAYLOAD = {
@@ -91,7 +91,7 @@ test('renders error message on short url error', async () => {
   render(
     <>
       <URLShortLinkButton {...props} />
-      <AntdToastContainer />
+      <ToastContainer />
     </>,
     { useRedux: true },
   );
