+++ superset-frontend/src/views/RootContextProviders.tsx
@@ -25,7 +25,7 @@ import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
 import getBootstrapData from 'src/utils/getBootstrapData';
 import { FlashProvider, DynamicPluginProvider } from 'src/components';
-import { NotificationProvider } from 'src/components/NotificationProvider';
+import { NotificationProvider } from 'src/components/MessageToasts/NotificationProvider';
 import { EmbeddedUiConfigProvider } from 'src/components/UiConfigContext';
 import { SupersetThemeProvider } from 'src/theme/ThemeProvider';
 import { ThemeController } from 'src/theme/ThemeController';
